Consolidate structured array into array.
