+++ tiled/_tests/test_structured_array.py
@@ -1,9 +1,9 @@
 import numpy as np
 import pytest
 
-from tiled.adapters.array import StructuredArrayTabularAdapter
+from tiled.adapters.array import ArrayAdapter
 from tiled.client import from_tree
-from tiled.structures.structured_array import StructDtype
+from tiled.structures.array import StructDtype
 from tiled.trees.in_memory import Tree
 
 
@@ -32,7 +32,7 @@ def test_read():
         [("Rex", 9, 81.0), ("Fido", 3, 27.0)],
         dtype=[("name", "U10"), ("age", "i4"), ("weight", "f4")],
     )
-    tree = Tree({"A": StructuredArrayTabularAdapter.from_array(data)})
+    tree = Tree({"A": ArrayAdapter.from_array(data)})
 
     client = from_tree(tree)
