+++ tiled/client/node.py
@@ -28,12 +28,6 @@ class Node(BaseClient, collections.abc.Mapping, IndexersMixin):
                 "array": lambda: importlib.import_module(
                     "..array", Node.__module__
                 ).ArrayClient,
-                "structured_array_generic": lambda: importlib.import_module(
-                    "..array", Node.__module__
-                ).ArrayClient,
-                "structured_array_tabular": lambda: importlib.import_module(
-                    "..array", Node.__module__
-                ).ArrayClient,
                 "dataframe": lambda: importlib.import_module(
                     "..dataframe", Node.__module__
                 ).DataFrameClient,
@@ -52,12 +46,6 @@ class Node(BaseClient, collections.abc.Mapping, IndexersMixin):
             {
                 "node": lambda: Node,
                 "array": lambda: importlib.import_module(
-                    "..array", Node.__module__
-                ).DaskArrayClient,
-                "structured_array_generic": lambda: importlib.import_module(
-                    "..array", Node.__module__
-                ).DaskArrayClient,
-                "structured_array_tabular": lambda: importlib.import_module(
                     "..array", Node.__module__
                 ).DaskArrayClient,
                 "dataframe": lambda: importlib.import_module(
