+++ tiled/server/core.py
@@ -354,8 +354,6 @@ def construct_entries_response(
 DEFAULT_MEDIA_TYPES = {
     "array": "application/octet-stream",
     "dataframe": APACHE_ARROW_FILE_MIME_TYPE,
-    "structured_array_tabular": "application/octet-stream",
-    "structured_array_generic": "application/octet-stream",
     "variable": "application/octet-stream",
     "data_array": "application/octet-stream",
     "dataset": "application/netcdf",
@@ -747,12 +745,6 @@ class WrongTypeForRoute(Exception):
 
 FULL_LINKS = {
     "array": {"full": "{base_url}array/full/{path}"},
-    "structured_array_generic": {
-        "full": "{base_url}structured_array_generic/full/{path}"
-    },
-    "structured_array_tabular": {
-        "full": "{base_url}structured_array_tabular/full/{path}"
-    },
     "dataframe": {"full": "{base_url}dataframe/full/{path}"},
     "variable": {"full": "{base_url}variable/full/{path}"},
     "data_array": {"full_variable": "{base_url}data_array/variable/full/{path}"},
