+++ tiled/server/models.py
@@ -42,8 +42,6 @@ class StructureFamilies(str, enum.Enum):
     node = "node"
     array = "array"
     dataframe = "dataframe"
-    structured_array_tabular = "structured_array_tabular"
-    structured_array_generic = "structured_array_generic"
     variable = "variable"
     data_array = "data_array"
     dataset = "dataset"
