+++ tiled/server/router.py
@@ -454,216 +454,6 @@ def array_full(
 
 
 @router.get(
-    "/structured_array_generic/block/{path:path}",
-    response_model=models.Response,
-    name="structured array (generic) block",
-)
-def structured_array_generic_block(
-    request: Request,
-    adapter=Depends(adapter),
-    block=Depends(block),
-    slice=Depends(slice_),
-    expected_shape=Depends(expected_shape),
-    format: Optional[str] = None,
-    serialization_registry=Depends(get_serialization_registry),
-):
-    """
-    Fetch a chunk of array-like data.
-    """
-    if adapter.structure_family != "structured_array_generic":
-        raise HTTPException(
-            status_code=404,
-            detail=f"Cannot read {adapter.structure_family} structure with /structured_array_generic/block route.",
-        )
-    if block == ():
-        # Handle special case of numpy scalar.
-        if adapter.macrostructure().shape != ():
-            raise HTTPException(
-                status_code=400,
-                detail=f"Requested scalar but shape is {adapter.macrostructure().shape}",
-            )
-        array = adapter.read()
-    else:
-        try:
-            array = adapter.read_block(block, slice=slice)
-        except IndexError:
-            raise HTTPException(status_code=400, detail="Block index out of range")
-        if (expected_shape is not None) and (expected_shape != array.shape):
-            raise HTTPException(
-                status_code=400,
-                detail=f"The expected_shape {expected_shape} does not match the actual shape {array.shape}",
-            )
-    try:
-        return construct_data_response(
-            "structured_array_generic",
-            serialization_registry,
-            array,
-            adapter.metadata,
-            request,
-            format,
-            expires=getattr(adapter, "content_stale_at", None),
-        )
-    except UnsupportedMediaTypes as err:
-        raise HTTPException(status_code=406, detail=err.args[0])
-
-
-@router.get(
-    "/structured_array_tabular/block/{path:path}",
-    response_model=models.Response,
-    name="structured array (tabular) block",
-)
-def structured_array_tabular_block(
-    request: Request,
-    adapter=Depends(adapter),
-    block=Depends(block),
-    slice=Depends(slice_),
-    expected_shape=Depends(expected_shape),
-    format: Optional[str] = None,
-    serialization_registry=Depends(get_serialization_registry),
-):
-    """
-    Fetch a chunk of array-like data.
-    """
-    if adapter.structure_family != "structured_array_tabular":
-        raise HTTPException(
-            status_code=404,
-            detail=f"Cannot read {adapter.structure_family} structure with /structured_array_tabular/block route.",
-        )
-    if block == ():
-        # Handle special case of numpy scalar.
-        if adapter.macrostructure().shape != ():
-            raise HTTPException(
-                status_code=400,
-                detail=f"Requested scalar but shape is {adapter.macrostructure().shape}",
-            )
-        array = adapter.read()
-    else:
-        try:
-            array = adapter.read_block(block, slice=slice)
-        except IndexError:
-            raise HTTPException(status_code=400, detail="Block index out of range")
-        if (expected_shape is not None) and (expected_shape != array.shape):
-            raise HTTPException(
-                status_code=400,
-                detail=f"The expected_shape {expected_shape} does not match the actual shape {array.shape}",
-            )
-    try:
-        return construct_data_response(
-            "structured_array_tabular",
-            serialization_registry,
-            array,
-            adapter.metadata,
-            request,
-            format,
-            expires=getattr(adapter, "content_stale_at", None),
-        )
-    except UnsupportedMediaTypes as err:
-        raise HTTPException(status_code=406, detail=err.args[0])
-
-
-@router.get(
-    "/structured_array_tabular/full/{path:path}",
-    response_model=models.Response,
-    name="structure array (tabular) full array",
-)
-def structured_array_tabular_full(
-    request: Request,
-    adapter=Depends(adapter),
-    slice=Depends(slice_),
-    expected_shape=Depends(expected_shape),
-    format: Optional[str] = None,
-    serialization_registry=Depends(get_serialization_registry),
-):
-    """
-    Fetch a slice of array-like data.
-    """
-    if adapter.structure_family != "structured_array_tabular":
-        raise HTTPException(
-            status_code=404,
-            detail=f"Cannot read {adapter.structure_family} structure with /structured_array_tabular/full route.",
-        )
-    # Deferred import because this is not a required dependency of the server
-    # for some use cases.
-    import numpy
-
-    try:
-        array = adapter.read()
-        if slice:
-            array = array[slice]
-        array = numpy.asarray(array)  # Force dask or PIMS or ... to do I/O.
-    except IndexError:
-        raise HTTPException(status_code=400, detail="Block index out of range")
-    if (expected_shape is not None) and (expected_shape != array.shape):
-        raise HTTPException(
-            status_code=400,
-            detail=f"The expected_shape {expected_shape} does not match the actual shape {array.shape}",
-        )
-    try:
-        return construct_data_response(
-            "structured_array_tabular",
-            serialization_registry,
-            array,
-            adapter.metadata,
-            request,
-            format,
-            expires=getattr(adapter, "content_stale_at", None),
-        )
-    except UnsupportedMediaTypes as err:
-        raise HTTPException(status_code=406, detail=err.args[0])
-
-
-@router.get(
-    "/structured_array_generic/full/{path:path}",
-    response_model=models.Response,
-    name="structured array (generic) full array",
-)
-def structured_array_generic_full(
-    request: Request,
-    adapter=Depends(adapter),
-    slice=Depends(slice_),
-    expected_shape=Depends(expected_shape),
-    format: Optional[str] = None,
-    serialization_registry=Depends(get_serialization_registry),
-):
-    """
-    Fetch a slice of array-like data.
-    """
-    if adapter.structure_family != "structured_array_generic":
-        raise HTTPException(
-            status_code=404,
-            detail=f"Cannot read {adapter.structure_family} structure with /structured_array_generic/full route.",
-        )
-    # Deferred import because this is not a required dependency of the server
-    # for some use cases.
-    import numpy
-
-    try:
-        array = adapter.read()
-        if slice:
-            array = array[slice]
-        array = numpy.asarray(array)  # Force dask or PIMS or ... to do I/O.
-    except IndexError:
-        raise HTTPException(status_code=400, detail="Block index out of range")
-    if (expected_shape is not None) and (expected_shape != array.shape):
-        raise HTTPException(
-            status_code=400,
-            detail=f"The expected_shape {expected_shape} does not match the actual shape {array.shape}",
-        )
-    try:
-        return construct_data_response(
-            "structured_array_generic",
-            serialization_registry,
-            array,
-            adapter.metadata,
-            request,
-            format,
-            expires=getattr(adapter, "content_stale_at", None),
-        )
-    except UnsupportedMediaTypes as err:
-        raise HTTPException(status_code=406, detail=err.args[0])
-
-
-@router.get(
     "/dataframe/partition/{path:path}",
     response_model=models.Response,
     name="dataframe partition",
